Moved Mod Scaffolding Functions into Separated Class
